templated OmegaHField2
