+++ src/pcms/adapter/omega_h/omega_h_field_layout.h
@@ -19,7 +19,7 @@ public:
                     int num_components, CoordinateSystem coordinate_system,
                     std::string global_id_name = "global");
 
-  std::unique_ptr<FieldT<Real>> CreateField() const override;
+  std::unique_ptr<FieldT<Real>> CreateFieldReal() const override;
 
   int GetNumComponents() const override;
   // nodes for standard lagrange FEM
