+++ src/pcms/field_layout.h
@@ -29,7 +29,7 @@ class FieldT;
 class FieldLayout
 {
 public:
-  virtual std::unique_ptr<FieldT<Real>> CreateField() const = 0;
+  virtual std::unique_ptr<FieldT<Real>> CreateFieldReal() const = 0;
 
   // number of components
   int virtual GetNumComponents() const = 0;
