+++ test/test_field_communication.cpp
@@ -71,7 +71,7 @@ void client1(MPI_Comm comm, Omega_h::Mesh& mesh, std::string comm_name,
   PCMS_ALWAYS_ASSERT(n == gids.size());
   Omega_h::parallel_for(n, OMEGA_H_LAMBDA(int i) { ids[i] = gids[i]; });
 
-  auto field = layout->CreateField();
+  auto field = layout->CreateFieldReal();
   field->SetDOFHolderData(pcms::make_const_array_view(ids));
 
   pcms::FieldLayoutCommunicator<pcms::Real> layout_comm(comm_name + "1", comm,
@@ -95,7 +95,7 @@ void client2(MPI_Comm comm, Omega_h::Mesh& mesh, std::string comm_name,
   auto gids = layout->GetGids();
   const auto n = layout->GetNumOwnedDofHolder();
 
-  auto field = layout->CreateField();
+  auto field = layout->CreateFieldReal();
   pcms::FieldLayoutCommunicator<pcms::Real> layout_comm(comm_name + "2", comm,
                                                         rdv, channel, *layout);
   pcms::FieldCommunicator2<pcms::Real> field_comm(layout_comm, *field);
@@ -151,7 +151,7 @@ void server(MPI_Comm comm, Omega_h::Mesh& mesh, std::string comm_name,
   Omega_h::Write<Real> ids(n);
   Omega_h::parallel_for(n, OMEGA_H_LAMBDA(int i) { ids[i] = 0; });
 
-  auto field = layout->CreateField();
+  auto field = layout->CreateFieldReal();
   pcms::FieldLayoutCommunicator<pcms::Real> layout_comm1(
     comm_name + "1", comm, rdv, channel1, *layout);
   pcms::FieldLayoutCommunicator<pcms::Real> layout_comm2(
